Create a macro for sharing field implementations.
